Python numba CUDA float64

+++ python/GemmDenseCUDA/GemmDenseCUDA64.py
@@ -55,11 +55,11 @@ def main():
 
     rng = np.random.default_rng()
     start = time.time()
-    A = rng.random((A_rows, A_cols), dtype=np.float32)
+    A = rng.random((A_rows, A_cols), dtype=np.float64)
     tmp = _print_time(start, "initialize A")
-    B = rng.random((B_rows, B_cols), dtype=np.float32)
+    B = rng.random((B_rows, B_cols), dtype=np.float64)
     tmp = _print_time(tmp, "initialize B")
-    C = np.zeros(dtype=np.float32, shape=(A_rows, B_cols))
+    C = np.zeros(dtype=np.float64, shape=(A_rows, B_cols))
     tmp = _print_time(tmp, "initialize C")
 
     grid_rows = ceil((A_rows + BLOCK_SIZE - 1) / BLOCK_SIZE)
